feat: 重构mainview
feat: 重构mainview
